Change StackTraceContainsLineNumber test
In CI the HttpClient stacktrace did not contain any frames with lineno != 0. We moved the stack trace capturing from the Span .ctor to the End method, since then the stack from http methods did not contain line numbers - only in CI. Turned on extra logging which shows that `frame?.GetFileLineNumber()` returns 0 for every frame in that case, so nothing we can do something about. Therefore simplified the test for a single manual sync span, which still aligned with the purpose of the test: we only wanna make sure we capture line numbers in at least 1 scenario - in that case `frame?.GetFileLineNumber()` seems to return real line numbers. So we go with that test.