Normalize transaction names for ASP.NET MVC Full Framework
This commit normalizes the transaction name for ASP.NET MVC
Full Framework to use the convention of

<HTTP method> <area>/<controller>/<action> [<param>*]

which aligns with the convention used in ASP.NET Core
transaction names. The transaction name is determined from
the RouteData values, and where values are present but the
response is a 404, unknown route is used.

Integration tests introduce an MVC area with a HomeController
and Index action, to assert that the area route datatoken is
taken into account. Without taking into account the area,
same named controller actions in areas would end up
aggregrated under the same transaction name.

Closes #201