Capture User id and email
This commit captures the User id and email of an
authenticated user in ASP.NET Full Framework.
Such values can be captured only when Claims
based identity is used and there is a ClaimsPrincipal,
which is used by ASP.NET Identity- see

https://docs.microsoft.com/en-us/aspnet/identity/overview/getting-started/introduction-to-aspnet-identity

ASP.NET Identity became the default user and membership
implementation for ASP.NET full framework back around
2013, supeseding the
Membership provider implementation introduced in
ASP.NET 2.0, that uses RolePrincipal (and FormsIdentity)
for an authenticated user, which have no built-in notion
of a user id or email.

Closes #540