Add ASP.NET Identity
This commit adds ASP.NET Identity to the
ASP.NET full framework sample. In including
this, the OWIN implementation to hook
up authentication middleware is also
introduced.

ASP.NET Identity uses Entity Framework
and the in-memory SQLite database to
create all the tables needed for application
users.

Introduce an AccountController to register
users, and allow logging in and out.
Registration confirmation and Password
reset emails are displayed as alerts on
subsequent pages, for convenience. A
real system would send emails :)

Some tidying up of views in line with the
move to Bootstrap 4.