Add test for captured user id and email
This commit adds an integration test to
assert that the user id and email are
captured for an authenticated user.

Use Anglesharp to drive actions, which
is closer to browser driven actions than
performing HTTP calls with HTTP client.
For example, it'll automatically handle
sending values for Anti-XSRF tokens
and cookies.