Grant read/execute to SQLite.Interop.dlls to IIS_USRS group (#979)
This commit adds access to read/execute to the IIS_USRS
group, which created ApplicationPoolIdentity accounts belong to.

When AspNetFullFramework.Tests are run with an application pool
configured to use an ApplicationPoolIdentity, the pool may have
insufficient access to execute the native SQLite.Interop.dlls used in
database tests running with SQLite. ApplicationPoolIdentity accounts
will not have access if the repository is cloned to and the application
run under IIS from a privileged location, such as %USERPROFILE%.

This commit fixes this issue by configuring read and
execute rights for the IIS_USRS
group on the SQLite.Interop.dll files when configuring the
ApplicationPool and using ApplicationPoolIdentity identity type.