Only initialize SampleDataDbContext database when required
This commit removes the instantion of SampleDataDbContext per
OWIN context and moves instantiation only to where it is needed.
First instantiation of SampleDataDbContext causes the in-memory
SQLite database to be created which in turn results in an APM
transaction and spans to be created, interfering with tests
that do not require SampleDataDbContext.

Simplify the Authentication and SignIn components down to
only what is needed.