Normalize transaction names for ASP.NET Full Framework (#973)
This commit normalizes the transaction name for ASP.NET MVC
Full Framework to use the convention of

<HTTP method> <area>/<controller>/<action> [<param>*]

which aligns with the convention used in ASP.NET Core
transaction names. The transaction name is determined from
the RouteData values.

- set unknown route only when the route data did not end up routing 
to a controller action. This is determined by looking for a 404 
HttpException coming from System.Web.Mvc.

- Add webforms page, routed webforms page, web api controller,
and area controller, to assert transaction name in each case.

- Integration tests introduce an MVC area with a HomeController
and Index action, to assert that the area route datatoken is
taken into account. Without taking into account the area,
same named controller actions in areas would end up
aggregrated under the same transaction name.

Closes #201