Add HttpContextCurrentExecutionSegmentsContainer (#992)
* Add HttpContextCurrentExecutionSegmentsContainer

This commit adds an implementation of ICurrentExecutionSegmentsContainer
that gets/sets the current transaction and span in HttpContext.Items,
in addition to AsyncLocal<T>, making them available within the
ExecutionContext of IIS.

AsyncLocal<T> are not propagated across the ASP.NET ExecutionContext
running in IIS, resulting in the current transaction set in Application_BeginRequest
returning as null from Agent.Tracer.CurrentTransaction in later subsequent
IIS events, when such events execute asynchronously. ASP.NET does
however propagate the current HttpContext through HttpContext.Current,
allowing values to be stored in HttpContext.Items hashtable and shared
across IIS events. The HttpContextCurrentExecutionSegmentsContainer
uses both AsyncLocal<T> and HttpContext.Items to get/set the current
transaction and span, allowing them to be available in both IIS
events and async contexts. AsyncLocal<T> is still needed for places
where HttpContext.Current is null, for example, DiagnosticSourceListeners
running on a separate thread.

It seems it may be possible for the current transaction and current
span instances stored in AsyncLocal<T> and HttpContext.Items to be different
in such scenarios where one is available but not the other on beginning
the transaction/span, and where availability is reversed on end. Unsure
how likely this is in practice however, and using both AsyncLocal<T> and
HttpContext.Items is the common approach in resolving this issue in < .NET 4.7.1.

For .NET 4.7.1, a new OnExecuteRequestStep method has been added to
HttpApplication, to allow the execution context to be restored/saved:

https://devblogs.microsoft.com/dotnet/net-framework-4-7-1-asp-net-and-configuration-features/#asp-net-execution-step-feature

Closes #934
Closes #972

* tidy after rebase

* Add concurrent requests test

This commit adds an integration test for
HttpContextCurrentExecutionSegmentsContainer
that makes multiple concurrent requests
and asserts that the captured transactions
and span ids align with expectations.