Add support for StackExchange.Redis
This commit adds support for tracing redis commands
sent with StackExchange.Redis.

The implementation registers a profiler with
IConnectionMultiplexer that provides a profiling session
per transaction. Added an Ended event to transaction
that a delegate is registered with to end the profiling session and
create a span for each profiled command.

Added StackExchangeRedisSample to demonstrate how to register
Elastic APM agent. For ASP.NET Core, registration is performed in
ConfigureServices(IServiceCollection services):

	services.AddSingleton<IConnectionMultiplexer>(provider =>
	{
		var connection = ConnectionMultiplexer.Connect("localhost");
		connection.UseElasticApm();
		return connection;
	});

Closes #874