AllowSynchronousIO when capturing request body
This commit allows synchronous IO for a given request
when capturing the request body.

From ASP.NET Core 3.0 onwards, synchronous reading or writing
of request streams is disabled by default, throwing
an exception when attempting to do so. See

https://docs.microsoft.com/en-us/aspnet/core/migration/22-to-30?view=aspnetcore-5.0&tabs=visual-studio#allowsynchronousio-disabled

Capturing of the request body by the agent is a synchronous
operation because it may happen in the the synchronous
flow of a diagnostic listener.

Consideration was made to using GetAwaiter().GetResult()
which would effectively be sync over async, but using the
http body control feature is preferred as it is a framework feature
designed for this purpose.

Fixes #1044.