Introduce ServerUrl and deprecate ServerUrls
This commit consolidates ServerUrl and ServerUrls options
by

- introducing ServerUrl configuration value
- deprecating ServerUrls with ObsoleteAttribute

ServerUrl uses the ServerUrl environment or setting key value,
falling back to ServerUrls[0] when not specified, allowing
the usage of ServerUrls to continue to work.

ServerUrls uses the ServerUrls environment or setting key value
if specified, falling back to ServerUrl environment or setting key value
when not specified, allowing exitsing usage of ServerUrls outside of the
agent API to work with ServerUrl.

Closes #1035