Add diff build target
This commit adds the ability to diff built assemblies against
the previously released version on Nuget. If there are changes,
an XML file containing the diffs is emitted in ./build/output
with a name matching the assembly name.

To run the diff command for all assemblies

```
build.bat diff
```

This will use the netstandard2.0 TFM by default, but a TFM
can be supplied with `--framework`.

To run for only certain assemblies

```
build.bat diff --packageids Elastic.Apm --packageids Elastic.SqlClient
```

Closes #1079