Update log level based on central configuration
This commit updates the log level of the configured
IApmLogger based on the log level received from
central configuration.

An ILogLevelSwitchable interface is introduced that
can be implemented by an IApmLogger to allow
the log level to be updated by central configuration.

ConsoleLogger implements ILogLevelSwitchable
but NetCoreLogger does not. The ability to change
the log level of NetCoreLogger at runtime is controlled
by the provider used with Microsoft.Extensions.Logging.

Closes #970