Wait for payload sender to queue events in tests (#1101)
This commit 

- introduces AutoResetEvent on MockPayloadSender
to block threads until events have been queued. An optional
timeout can be supplied to time out the waiting for an event.

- Removes SerializerMockPayloadSender.
It was used to ensure that fields were sanitized at
serialization time. Now that sanitization happens in
filters, there is no need to serialize in sender tests anymore.

- Adds Signal methods to MockPayloadSender
that can be called in scenarios that assert the absence of
event, and the order of execution is known, to prevent
having to wait for a specified timeout.

- Remove stages timeout from Jenkinsfile
- Fix an issue where the Deep copying of Response was not assigned