Ensure ETW session is deleted when GcMetricsProvider is disposed
This commit fixes a bug in collecting GC Metrics on
.NET Framework where the underlying ETW session created by
TraceEventSession was not stopped and deleted when
GcMetricsProvider was disposed. The issue arises due
to the using around _traceEventSession in the
GcMetricsProvider's ctor, and the Task.Run() inside it,
which leaves the session in a running state that can't
be deleted on disposing the GcMetricsProvider.

An internal TraceEventSession name property exposes
the name of the session to allow tests to assert that
the session is deleted on dispose.

Fixes #897