Artifact dotnet hooks zip file
This commit updates the build project pack target to
build nuget packages and the zip file for startup hooks.

The CI linux scripts are updated to use the build project
pack target, requiring them to look in one directory
for nuget packages, and allowing them to artifact the
zip file.