Propagate sample rate through tracestate
This commit propagates the sample rate of the
APM agent through the tracestate header, and
uses the sample rate received in incoming tracestate
as the sample rate for non-root transactions and spans.

Introduce TraceState type to model tracestate,
allowing incoming headers to be added and sample
rate to be set. Move TraceState validation from
TraceContext onto TraceState.

Obsolete TryDeserializeFromString and SerializeToString
on DistributedTracingData in favour of a method that accepts
both traceparent and tracestate, and properties that get
the traceparent and tracestate text headers, respectively.

Change Transaction and Span SampleRate to double?
to allow a value to be omitted when incoming
Trace context does not contain a tracestate header
or the header does not contain the es vendor s attribute.

Closes #1021