StartupHook fix: remove check on Microsoft.AspNetCore assemblies (#1165)
This commit removes the check for a loaded assembly starting with
Microsoft.AspNetCore.* when loading DiagnosticSubscribers related
to ASP.NET Core. For some applications, such as a Web API on
netcoreapp3.1, the assemblies are loaded after the startup hook runs,
so the APM agent doesn't end up subscribing to the DiagnosticSource
related to ASP.NET Core. 

Co-authored-by: Russ Cam <russ.cam@elastic.co>