Test startup hooks (#1167)
This commit tests that the APM agent startup hook feature is able to
configure and setup the agent for ASP.NET Core 3.0, 3.1 and
ASP.NET 5, and send APM data to APM server.

A sample ASP.NET (Core) application is included as a target application
for startup hooks. Tests start the application, along with a test
APM server, make a request to the application, and assert that
APM data is captured by APM server.

A DotnetProject type is introduced that executes commands with
dotnet executable to create, publish and run dotnet projects. This is
used to test startup hooks for webapi, webapp and mvc projects for
netcoreapp3.0, netcoreapp3.1 and net5.0 frameworks.

Startup hook tests run as part of the Tests stage on Linux.
A separate Startup Hooks stage is introduced on Windows for testing.

Other changes:
- Loosen agent version assertion in Test APM server

  The Test APM server may be run with an agent version built from a different commit
  than the current one, so loosen the agent version assertion to just check for 
  major.minor.patch _without_ the commit suffix.
- Add OnReceive event to Test APM server

  Allows blocking a thread until APM server receives data, which is useful for testing.

- Don't invoke MSBuild on CI in build project automation

  Not needed for the current setup. F# Make (FAKE) appears to be invoking MSBuild.exe
  that does not have access to .NET Core SDKs, which requires further investigation.

- Add dotnet 3.0 SDK to CI tests

  Added to build and test startup hooks for netcoreapp3.0
 
Closes #1073