Add support for self signed certs (#1180)
This commit adds support for communicating with APM server
that has been configured with a certificate to use for
SSL/TLS that is not trusted e.g. self-signed certificates
or certificates issued by a Certificate Authority (CA) not in the trust store.

A path can be specified with `ElasticApm:ServerCert` to
a PEM encoded certificate used to secure APM server. A
`ServerCertificateCustomValidationCallback` is configured
on `HttpClientHandler` of the client used to communicate with
APM server that validates the certificate through public
key pinning.

Add integration tests for ServerCert and VerifyServerCert.
Extend Mock APM server to support HTTPS configuration
with a specified certificate.

Closes #1171