Runs tests in blame mode (#1193)
This commit runs tests in blame mode with a blame-hang
timeout of 5 minutes. The test sequence and blame dump files
are artifacted if there is a hang.

- Add xUnit attribute to capture and restore the ActivityDefaultIdFormat
for tests that change Activity.DefaultIdFormat static property. Introduce
a test collection for these tests so that they are run sequentially.

- Use dotnet test for testnet461.bat file instead of vstest. This allows the use
of blame-hang arguments.

- Update Visual Studio choco packages to newer version.

- Use agent in tests with a using statement to dispose at test end.

- Increase log level change detection test to 60 seconds as 20 seconds is
  not always enough.

- Introduce default timeout in mock payload sender when waiting for events.
  This is to prevent waiting forever and hanging. As soon as events are received,
  waiting threads are unblocked.

- Dispose HttpListener when attempt to start throws Exception when creating
  LocalServer.

- Add XUnitLogger to capture test output with XUnit.

(cherry picked from commit 77ce7870551439a3859d5b98cca78224bc5db1f7)