Skip HttpCallWithW3CActivityFormat test on NET461 (#1203)
This commit skips the HttpCallWithW3CActivityFormat for .NET 461.  This test can intermittently fail on .NET 461 due to the way in which HTTP calls with HttpWebRequest are instrumented, and the concurrent running of tests; Interaction with ServicePoint causes race condition in the registration of a ServicePoint object with ServicePointManager, resulting in the HTTP request not being tracked by the Diagnostic Source, and hence no Span captured for the call.