address PR feedback
This commit removes the Obsolete attributes from DistributedTracingData,
and does not expose Tracestate for now. Can evaluate again in future.

When an activity is created when a transaction is started, use the tracestate
from the activity, if present, but set the sample rate from the sampler and
mutate tracestate if the es vendor is present. A transaction created without
explicitly passed DistributedTracingData is considered to be a root transaction,
so the sample rate is governed by the sampler.