Propagate sample rate through tracestate (#1147)
This commit propagates the sample rate of the
APM agent through the tracestate header, and
uses the sample rate received in incoming tracestate
as the sample rate for non-root transactions and spans.

Introduce TraceState type to model tracestate,
allowing incoming headers to be added and sample
rate to be set. Move TraceState validation from
TraceContext onto TraceState.

Change Transaction and Span SampleRate to double?
to allow a value to be omitted when incoming
Trace context does not contain a tracestate header
or the header does not contain the es vendor s attribute for
a non-root transaction.

When an activity is created when a transaction is started, use the tracestate
from the activity, if present, but set the sample rate from the sampler and
mutate tracestate if the es vendor is present. A transaction created without
explicitly passed DistributedTracingData is considered to be a root transaction,
so the sample rate is governed by the sampler.

Closes #1021