Add instrumentation for Azure Service Bus (#1225)
This commit adds instrumentation for Azure Service Bus when an application is 
using Microsoft.Azure.ServiceBus 3.0.0+ or Azure.Messaging.ServiceBus 7.0.0+ nuget packages.

Two IDiagnosticListener implementations, one for Microsoft.Azure.ServiceBus 
and another for Azure.Messaging.ServiceBus, create transactions and spans for received 
and sent messages:

A new transaction is created when

- one or more messages are received from a queue or topic subscription.
- a message is receive deferred from a queue or topic subscription.

A new span is created when there is a current transaction, and when

- one or more messages are sent to a queue or topic.
- one or more messages are scheduled to a queue or a topic.

The diagnostic events do not expose details about sent or received messages.
The trace ids of messages are exposed but are not currently captured in this implementation.
Messages are often received in batches, and it is possible for each message to have its
own trace id, but the APM implementation does not have a concept for capturing such
data right now. See https://github.com/elastic/apm/issues/122

A terraform template file is used to create a resource group, Azure Service Bus namespace 
resource in the resource group, and set RBAC rules to allow the Service Principal that issues
the creation access to the resources. The Service Principal credentials can are sourced from
a .credentials.json file in the root of the repository for CI, and from an account authenticated
with az for local development. A default location is set within the template, but all variables 
can be passed using standard Terraform input variable conventions.

Closes #1157