Use TraceLogger as default logger in ASP.NET Full Framework
This commit adds a logger implementation, TraceLogger, that
writes agent logs to a TraceSource with name "Elastic.Apm".
This logger is configured as the default logger for
ASP.NET Full Framework applications, which can use configuration
to write log messages to file, debug window, Windows event log, etc.

Add a section to docs with an example of how to configure
the logger in web.config.

Move the default log level from ConsoleLogger into DefaultValues.

Closes #1263