Create HTTP spans for known services (#1258)
This commit adds the ability to register IHttpSpanTracers
to be able to create HTTP spans for known services.

The HttpDiagnosticListener enumerates registered tracers
passing details about the HTTP request, which a tracer
can determine whether it is a match for a known service.
If it is a match, the tracer has the opportunity to create a
span. If none of the tracers determine it to be a match,
the HttpDiagnosticListener may create a "normal" HTTP span
for the HTTP request.

Since IHttpSpanTracers are dependent on a subscribed
HttpDiagnosticListener, adding a tracer checks to see if one
is subscribed, and subscribes one if not. When the HttpDiagnosticListener
is subscribed by a tracer, it won't create "normal" HTTP spans
for a HTTP request if none of the tracers determine it to be a match.
Explicitly subscribing HttpDiagnosticListener will create "normal"
HTTP spans, or in the event that a HttpDiagnosticListener is
already subscribed, will set configuration to create "normal" HTTP spans
when a tracer does not match. Such an implementation doesn't expose 
the implementation details of span capture to a user.