Do not capture HTTP child spans for Elasticsearch
This commit updates the Elasticsearch integration to
not capture HTTP child spans for calls to Elasticsearch.

An instrumentation flag is set by spans started by the
Elasticsearch integration diagnostic listeners that is
checked in the HttpDiagnosticListener.

We might consider refactoring to include an
Exit property on span in the future, that can be checked.

Closes #1276