Better logging in PayloadSenderV2 on task cancellation
This commit updates PayloadSenderV2 to handle task cancellation
exceptions specifically, that may occur during the sending workloop
iteration.

The number of items that were not sent is logged, but the exception
no longer is, as it doesn't provide value. In addition, the exception
is rethrown to allow BackendCommComponentBase to log cancellation.

- Move Metadata to Model
- Remove use of Indent() extension method and replace with indentation
in the joining characters. Indent the first item.