Log managed integrations to separate log file (#44)
This commit changes the logging for managed integrations to log to
a log file in the profiler log directory. Previously, managed integrations
used the agent logger, however, the agent singleton instance may not
be instantiated at the point a logger is needed e.g. when deferring
agent instantiation to the ElasticApmModule for ASP.NET integration.