Improve Performance counter handling (#1536)
This commit improves the handling of performance counters
used to get the total CPU % time on Windows. When
the "Processor" performance counter category does not exist,
try using the "Processor Information" category name.

Include information in the log message about checking permissions
only when an UnauthorizedAccessException is thrown. Add
link to documentation for more details.

Closes #1505