Set span context destination service name and type to empty string
This commit updates the Span context destination service name
and type to set them to empty string by default.

Both fields are marked as obsolete and will be removed in a
future version, but need to be sent as empty strings for
older APM server compatibility, which validates the presence
of these fields.

Fixes #1563