Change OTel version check logic (#1648)
During startup, the APM Server version is not known. This PR introduces a callback which is triggered once the agent fetches the server version.

The bridge is enabled with unknown server version (in order to not miss early spans), and then a final APM Server version check is done in the callback with proper logging.