Update Jenkinsfile
Removing `Opbeans` stage from CI.

What's this stage: we run this when we release. The intention with it is to update opbeans-dotnet to the agent version we release in the same pipeline.

Why we remove: unfortunately this consistently fails, because we try to update the agent packages in opbeans-dotnet to a version which we just pushed a few seconds earlier in an earlier step. This never works, because NuGet needs some time to index the new versions. Therefore we need another way to automatically bump up the agent version in opbeans-dotnet.