Enable CloudMetadataProvider on Azure Functions
Make sure that setting cloud meta data also works in Azure Functions.
There are subtle difference to the existing `AzureAppServiceMetadataProvider`, which is why a dedicated provider class was implemented for that. However, a lot of the existing code could be reused and was refactored into the newly introduced base class `EnvironmentBasedAzureMetadataProvider`.