Remove ETW powered GC metrics on FullFramework
This requires more than the documented `Performance Monitor Users`
privileges.

To read from ETW the user also needs to be part of the `Performance Log
Users` group.

However to start a new ETW session the user needs to be LocalSystem,
LocalService, NetworkService or part of the Local Admin group.

This goes against IIS best practices and not something we feel
comfortable telling our users to run their Application Pools under.

Since IIS is a big deployment target for full framework and errors
related to this metric collection come up alot in this context we opt to
pulling this feature out and making it a .NET core only feature.

Currently this metric is not used in the APM UI.