No longer parse request cookies, but ensure they are still redacted in the `Cookie` header string (#2444)
We've identified that sending transactions with cookies parsed to
`Request.Cookies` can result in failures to index those transactions
when any of the cookies include periods (which is common for ASP.NET
core by default).

This PR removes the extraction from the header onto `Request.Cookies`.
Instead, we now modify the `Cookie` header value to redact the cookie
items based on the configured list of field names to sanitize. For .NET
targets, we optimize this to avoid allocations as much as possible.
Unfortunately, for the .NET Framework, we can't be as efficient.