Fixes and enhancements for Azure Functions (isolated worker) (#2505)
This PR fixes several things in Azure Functions (isolated worker) which
occur when using the latest templates for Azure Functions.

- Newer versions of the Functions library prefer the gRPC-based
implementation, which throws when `Url` is accessed. Instead, we read
these from other data on the `FunctionContext`.
- Address Functions library changes that break distributed tracing. We
now parse the original headers from the `BindingContext` instead of the
request, which may contain a `traceparent` with the sampling flag set to
false when the user request does not include a specific `traceparent`.
- We explicitly don't record activities from the Azure functions library
as these are pretty broken
(https://github.com/Azure/azure-functions-dotnet-worker/issues/2733) and
are redundant when using our middleware.
- Downgrade several packages as the newer ones are now deprecated
(thanks for the confusion, Microsoft!).
- Update some outdated compiler pre-processor directives.
- The final few commits focus on CI integration test hangs on Linux. We
don't have a perfect solution for those, but after reviewing the hang
dumps, I've avoided some of the potential causes of the hangs. We'll
monitor subsequent PRs, and if they remain stable, we will readdress the
original causes.

A follow-up PR will update our documentation.

Closes #2407 
Closes #2311
Closes #2218