Avoid NullReferenceException on 401 Unauthorized (#2561)
We frequently experience NullReferenceExceptions in our logging
registered by the first chance exception handler.

It is caused by `intakeResponse` being not null: Accepted = 0, Errors =
null, and then evaluating `intakeResponse.Errors.Count` triggering an
error.

Environment:

Unauthorized 401 on
https://53b082e7cfeb4d9a9c117371xxxxxxxx.apm.eu-west-1.aws.cloud.es.io/intake/v2/events

The payload returned is:

```json
{"error":"authentication failed"}\n
```

So an improvement might be to also deserialize this JSON format.

Co-authored-by: Steve Gordon <sgordon@hotmail.co.uk>