Add traceStateValue parameter to DistributedTracingData.TryDeserializ… (#2607)
We pass the 'traceparent' value in our asynchronous message
communication.
We follow the instructions at
https://www.elastic.co/guide/en/apm/get-started/7.10/distributed-tracing.html#_manual_distributed_tracing.
However, it is not possible to restore the 'tracestate' as it is for the
Go agent implementation.

In pair 'restart_external' TraceContinuationStrategy, it is not possible
to search the logs across services using the original traceid, as each
message handling generates anew traceid due to the absence of a
tracestate.

The PR suggests a possible solution.