Remove post release step to backport docs to 1.x (#2644)
As titled as discussed with the docs team.

This step is failing and needs work to fix it or manually backport. We
expect the docs on main to be the source of truth or all 1.x, so for
now, don't need to maintain updating old asciidocs.

cc @theletterf / @colleenmcginnis  / @Mpdreamz