Add `netstandard2.1` target and update `DiagnosticSource` dependency (#2679)
Fixes #2657
Fixes #2445

This adds a `netstandard2.1` target so that features such as OTel bridge
are available on older runtimes, even though those are out of support.
This prevents an update of our package, without a runtime update,
breaking/removing features. Features may still not be 100% compatible
with older unsupported runtimes, but should work in general.

As part of this, I updated the minimum dependency for
`System.Diagnostics.DiagnosticSource` which is no generally 8.0.0. We
still ship a version of `Elastic.Apm` and loader for the Startup hooks
zip file which are built against 6.0.0. These support older unsupported
runtimes although with no official testing.

NOTE: We DOT NOT update `System.Diagnostics.DiagnosticSource` on
`net462` and `net472`. While this worked and tested fine in local dev,
it fails the IIS tests in CI. For now, we leave the version in these
targets unchanged from 5.0.0. This is likely fine as we do not depend on
any newer APIs in those targets.

Compiler pre-processor directives now prefer `NET` over a specific .NET
version. We also add `NETSTANDARD2_1` so that features continue to work
on older runtimes without a specific .NET target, i.e. 6 and 7.

Docs PR to follow.